backend/requirements.txt
flask>=2.3.0
flask-cors>=3.0.10
torch==2.2.0
torchvision==0.17.0
opencv-python==4.7.0.72
numpy>=1.26.0
Pillow==9.5.0
ultralytics==8.0.196 